refactor: v1 컨트롤러와 v2 컨트롤러 패키지 정리
