+++ app/src/main/java/com/techcourse/controller/v1/LogoutController.java
@@ -1,5 +1,6 @@
-package com.techcourse.controller;
+package com.techcourse.controller.v1;
 
+import com.techcourse.controller.common.UserSession;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
