+++ app/src/main/java/com/techcourse/controller/v1/RegisterController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse.controller.v1;
 
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
