+++ app/src/main/java/com/techcourse/controller/v2/LoginViewController.java
@@ -1,10 +1,11 @@
-package com.techcourse.annotationController;
+package com.techcourse.controller.v2;
 
 import com.interface21.context.stereotype.Controller;
 import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.view.JspView;
+import com.techcourse.controller.common.UserSession;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
